requirements.txt
Flask
requests
beautifulsoup4
mysql-connector-python
flask_login